part4 Exercises 4:20 - 4:23

+++ part4/blog/backend/controllers/blog.js
@@ -4,12 +4,30 @@ const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
 
-const getTokenFrom = (request) => {
+// 4.20*: bloglist expansion, step8
+// Middleware function to authenticate user
+const tokenExtractor = async (request, response, next) => {
     const authorization = request.get("authorization");
-    if (authorization && authorization.startsWith("Bearer ")) {
-        return authorization.replace("Bearer ", "");
+    let token = null;
+    if (authorization?.toLowerCase().startsWith("bearer ")) {
+        token = authorization.substring(7);
     }
-    return null;
+
+    // eslint-disable-next-line no-undef
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: "Unauthorized access" });
+    }
+
+    const user = await User.findById(decodedToken.id);
+
+    if (!user) {
+        return response.status(401).json({ error: "Unauthorized access" });
+    }
+
+    request.user = user;
+    next();
 };
 
 blogRouter.get("/", (request, response) => {
@@ -40,12 +58,26 @@ blogRouter.put("/api/blogs/:id", (request, response, next) => {
         .catch((error) => next(error));
 });
 
-blogRouter.delete(`/api/blogs/:id`, async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
-});
+blogRouter.delete(
+    `/api/blogs/:id`,
+    tokenExtractor,
+    async (request, response) => {
+        const blog = await Blog.findByIdAndRemove(request.params.id);
+        if (!blog) {
+            return response.status(404).json({ error: "Blog not found" });
+        }
 
-blogRouter.post("/api/blogs", async (request, response) => {
+        if (blog.user.toString() !== request.user.id.toString()) {
+            return response.status(401).json({ error: "Unauthorized access" });
+        }
+
+        await Blog.findByIdAndRemove(request.params.id);
+
+        response.status(204).end();
+    }
+);
+
+blogRouter.post("/api/blogs", tokenExtractor, async (request, response) => {
     const { title, url, likes, author } = request.body;
 
     if (!title) {
@@ -55,17 +87,7 @@ blogRouter.post("/api/blogs", async (request, response) => {
     if (!url) {
         return response.status(400).json({ error: "url is missing" });
     }
-
-    // eslint-disable-next-line no-undef
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
-
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: "token invalid" });
-    }
-
-    const user = await User.findById(decodedToken.id);
-
-    // const blog = new Blog(request.body);
+    const user = request.user;
     const blog = new Blog({
         title: title,
         author: author,
